Fixed a bug in the wem library (NaN in precond.c) and fixed unit conversion problem.
Note: the unit conversion shoulbe done in a better way. Now it is such that
I get all the way through same numbers as in the old code for ease of comparison.
This makes units quite messy for now. AU should be used throughout and other units ONLY for I/O
in case the user wants so.