Host writer function is passed as a function pointer
The naive way of setting the function for flushing library output to the
host was found to be problematic on Windows and Mac OS X yielding
undefined symbols at the library linking stage.
Using a function pointer gives greater flexibility and avoids this
non-cross-platform behaviour.
The signature for the function has been simplified to accept only one
argument, i.e. the output to be printed.
For Fortran hosts, the length can in fact be computed on-the-fly,
without the need to call the `pcmsolver_c2f_string` function.