Add Danger.Systems integration to the project
This commit set ups integration of the project with Danger.Systems, in
order to automatize some common reviewing tasks.
Danger.Systems is a Ruby gem that is installed and run in one the Travis
CI jobs and checks PRs for compliance with a set of rules defined in the
Dangerfile.
A bot is configured to report the results of the following checks for
any PR:
1. Commit linting
2. Additions to documentation and tests
3. Additions to CHANGELOG.md
4. Compliance to the code style

Tasks accomplished in this commit:
- The .travis.yml file was updated to run the Danger gem
- The .clang-format file was regenerated based on clang-format v3.9.1
- The contributing guidelines in CONTRIBUTING.md now mention the
  comments that will be posted to the PR by the bot
- Documentation about maintenance was updated to describe the role of
  Danger.Systems and the usage of Git pre-commit hooks
- The shebang in the shell scripts used as Git pre-commit hooks was fixed
- The license_maintainer.py script was adjusted in order to be Python 3 compatible.
- Danger.Systems uses v0.1.0 of code style validation plugin.
  The plugin comes from the official, upstream repository.
  Thanks @nikolaykasyanov!!!