Write to temporary file uppercased input, fix #91
The input file is uppercased, to ensure a consistent case is
used throughout. However, the original input file as provided by the
user, was being overwritten by its uppercased counterpart.
This annoying bug was noted by @ilfreddy.
The current commit fixes #91 by writing the uppercased text to
a temporary file. The temporary file is removed as soon as the parsing
has been done.